Get rid of MEMonitorIInventory
